Implements enums, view changes, total sum calculation

+++ src/DeleteExpenseFunction.java
@@ -10,25 +10,33 @@ public class DeleteExpenseFunction implements ActionListener {
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        Expense selectedExpense;
-        int rowIndex = mainScreen.table.getSelectedRow();
-        if (!mainScreen.filteredList.isEmpty()) {
-            selectedExpense = mainScreen.filteredList.get(rowIndex);
-        } else {
-            selectedExpense = ExpenseRepository.dataList.get(rowIndex);
+        if (mainScreen.state == State.MAIN) {
+            Expense selectedExpense;
+            int rowIndex = mainScreen.table.getSelectedRow();
+            if (!mainScreen.filteredList.isEmpty()) {
+                selectedExpense = mainScreen.filteredList.get(rowIndex);
+            } else {
+                selectedExpense = ExpenseRepository.dataList.get(rowIndex);
+            }
+            if (!mainScreen.filteredList.isEmpty()) {
+                mainScreen.filteredList.remove(selectedExpense);
+                ExpenseRepository.removeExpense(selectedExpense);
+            } else {
+                ExpenseRepository.removeExpense(selectedExpense);
+
+            }
+
+            mainScreen.centerPanel.removeAll();
+            if (!mainScreen.filteredList.isEmpty()) {
+                mainScreen.drawMainCenterPanel(mainScreen.filteredList);
+
+            } else {
+                mainScreen.drawMainCenterPanel(ExpenseRepository.dataList);
+            }
+
+            mainScreen.UI.jframe.revalidate();
+            mainScreen.UI.jframe.repaint();
         }
-        if (!mainScreen.filteredList.isEmpty()) {
-            mainScreen.filteredList.remove(selectedExpense);
-            ExpenseRepository.removeExpense(selectedExpense);
-        } else {
-            ExpenseRepository.removeExpense(selectedExpense);
 
-        }
-
-        mainScreen.centerPanel.removeAll();
-        mainScreen.drawCenterPanel(mainScreen.filteredList);
-
-        mainScreen.UI.jframe.revalidate();
-        mainScreen.UI.jframe.repaint();
     }
 }
